requirements.txt
tensorflow
flask
numpy
pillow
keras